+++ system/info_windows.go
@@ -9,6 +9,12 @@ package system
 //                                                                                    //
 // ////////////////////////////////////////////////////////////////////////////////// //
 
+import (
+	"time"
+)
+
+// ////////////////////////////////////////////////////////////////////////////////// //
+
 // LoadAvg contains information about average system load
 type LoadAvg struct {
 	Min1  float64 `json:"min1"`  // LA in last 1 minute
@@ -136,7 +142,7 @@ func GetNetworkSpeed(duration time.Duration) (uint64, uint64, error) {
 // CalculateNetworkSpeed calculate network input/output speed in bytes per second for
 // all network interfaces
 func CalculateNetworkSpeed(ii1, ii2 map[string]*InterfaceInfo, duration time.Duration) (uint64, uint64) {
-	return 0, 0, nil
+	return 0, 0
 }
 
 // GetIOUtil return IO utilization
@@ -146,7 +152,7 @@ func GetIOUtil(duration time.Duration) (map[string]float64, error) {
 
 // CalculateIOUtil calculate IO utilization for all devices
 func CalculateIOUtil(ci1 *CPUInfo, fi1 map[string]*FSInfo, ci2 *CPUInfo, fi2 map[string]*FSInfo) map[string]float64 {
-	return map[string]float64{"/": 0}, nil
+	return map[string]float64{"/": 0}
 }
 
 // ////////////////////////////////////////////////////////////////////////////////// //