+++ changelog.md
@@ -1,5 +1,11 @@
 ## Changelog
 
+### v7.5.0
+
+* `[errutil]` Implemented error interface (_added method_ `Error() string`)
+* `[errutil]` Minor improvements
+* `[system]` Fixed windows stubs
+
 ### v7.4.0
 
 * `[fmtutil]` Added flag `SeparatorFullscreen` which enable full size separator